Added an environment variable that indicates non-interactive mode

+++ bootstrap.sh
@@ -63,11 +63,16 @@ ln -s $HOME/.dotfiles/emacs.d $HOME/.emacs.d
 echo Installing tmux settings
 ln -s $HOME/.dotfiles/tmux/tmux.conf $HOME/.tmux.conf
 
-while true; do
-    read -p "Would you like to install the modified keyboard layout for Russian keyboards? [y/n] " yn
-    case $yn in
-        [Yy]* ) echo Installing modified Xmodmap; ln -s $HOME/.dotfiles/linux/Xmodmap $HOME/.Xmodmap; break;;
-        [Nn]* ) exit;;
-        * ) echo "Please answer yes or no.";;
-    esac
+if [ "$BATCH_MODE" == 1 ]; then
+    exit 0
+else
+    while true; do
+	read -p "Would you like to install the modified keyboard layout for Russian keyboards? [y/n] " yn
+	case $yn in
+            [Yy]* ) echo Installing modified Xmodmap; ln -s $HOME/.dotfiles/linux/Xmodmap $HOME/.Xmodmap; break;;
+            [Nn]* ) exit;;
+            * ) echo "Please answer yes or no.";;
+	esac
 done
+fi
+
